requirements.txt
Flask==2.1.1
pandas==1.4.2
plyer==2.0.0
pyserial==3.5
dbus-python==1.3.2